requirements.txt
paddlepaddle
paddleocr
pytesseract
opencv-python
pillow
numpy